Commentaire massif sur graphicelement, ajout du splashscreen

+++ runner/balle.h
@@ -4,8 +4,8 @@
 class Balle
 {
 private:
-    int m_x;
-    int m_y;
+    float m_x;
+    float m_y;
     int m_h;
     int m_w;
     int m_dx;
@@ -14,8 +14,8 @@ private:
 public:
     Balle(int x, int y, int w, int h, int dx, int dy);
     void move();
-    int getX() const;
-    int getY() const;
+    float getX() const;
+    float getY() const;
     int getH() const;
     int getW() const;
     void setDX(int &b);
